+++ ext/src/provides/dialect/SQLite/SQLiteExpress.cs
@@ -40,6 +40,9 @@ namespace mooSQL.data
         public override string charIndex(string substring, string str, string start)
             => $"(CASE WHEN {start} < 1 THEN 0 ELSE INSTR(SUBSTR({str}, {start}), {substring}) + {start} - 1 END)";
 
+        public override string isNullOrWhiteSpace(string expr)
+            => $"({expr} IS NULL OR TRIM({expr}) = '')";
+
         public override string dateDiffHour(string start, string end)
             => $"round((julianday({end}) - julianday({start})) * 24)";